+++ src/modules/BlocksData.py
@@ -6,9 +6,9 @@ import xml.etree.ElementTree as ET
 
 from src.core import logs
 from src.modules.BlocksDatabase import Database
-from src.modules.XML.ProgramBlocks import VariableSection
-from src.modules.XML.ProgramBlocks import Base, PlcEnum
-from src.modules.PlcBlocks import import_xml_to_block_group
+from src.modules.ProgramBlocks import VariableSection
+from src.modules.ProgramBlocks import Base, PlcEnum
+from src.modules.ProgramBlocks import import_xml_to_block_group
 
 logs.setup(logging.DEBUG)
 logger = logging.getLogger(__name__)
